+++ CHANGELOG.md
@@ -3,6 +3,8 @@ CRUDlex Changelog
 
 ## 0.9.7
 Released: Upcoming
+- CRUDData::listEntries() can take now operators for the filter parameter
+- Added filters for the list views
 
 ## 0.9.6
 Released: 2015-02-20